+++ src/capture/client/Capture/Output/FileWriter/FileWriter.h
@@ -28,7 +28,7 @@
 - (void)initializeWriter;
 - (NSString *)getAbsolutePath:(NSString *)relativePath;
 - (void)createDirectory:(NSString *)relativePath;
-- (void)writeData:(NSData *)data relativePath:(NSString *)relativePath timestamp:(uint64_t)timestamp;
+- (void)writeData:(NSData *)data relativePath:(NSString *)relativePath timestamp:(Float64)timestamp;
 
 - (void)handleTrash;
